+++ R/zzz.R
@@ -97,9 +97,8 @@ move_env_elements(Expr, pl, c("lit"), remove = FALSE)
 
 
 # pl$[readers]
-pl$scan_csv_pl = scan_csv_pl
-pl$csv_reader = csv_reader
-pl$read_csv = read_csv_
+pl$scan_csv = scan_csv
+pl$read_csv = read_csv
 pl$scan_arrow_ipc = scan_arrow_ipc